claude-computer-use-tutorial/pyproject.toml
[project]
name = "claude-computer-use-tutorial"
version = "0.1.0"
description = "Claude Computer Use agent examples from techyowls.io"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
